Tweak waveform and pathrender. Cleanup gl utility code

+++ src/gl/gl_vbo.h
@@ -5,12 +5,12 @@ struct DrawVBO {
     uint32_t vaoId      = 0;
     uint32_t vboVertId  = 0;
     uint32_t vboIdxId   = 0;
-    int64_t nIndices   = 0;
-    int64_t vboVertSize = 0;
-    int64_t vboIdxSize  = 0;
+    uint32_t nIndices   = 0;
+    size_t vboVertSize = 0;
+    size_t vboIdxSize  = 0;
     ~DrawVBO();
     void destroy();
     void genBuffers();
 
-    void uploadBuffer(uint32_t bufferType, void* ptr, int64_t len);
+    void uploadBuffer(uint32_t bufferType, void* ptr, size_t len);
 };
